Added section header type SHT_OPENBSD - value 0xffffffff, section header name ELF_OPENBSD - value ".openbsd", and elf_hash() prototype in preparation for Per's import of the ELF RTLD.
Every OpenBSD ELF bin/so should contain a "empty" section header of
type SHT_OPENBSD, and this string ".openbsd" in the .dynstr section.
This will allow us to determine the exe/lib's target OS.  Note that
use of SHT_OPENBSD and ELF_OPENBSD is not ABI compliant.  GNU binutils
will require mods to met this goal.

The relocation settings are now re-enabled.