./pf_key_v2.c: Merge with EOM 1.11
exchange_establish do checking of existing exchanges these days

Remove SA_FLAG_REPLACED settings from various parts in preparation of a
grand unified setting in exchange_finalize.  Fix sa_mark_replaced to not
release a referance to the sa, and adjust the API as it won't get called
as a finalize func anymore.