add some ifmedia goop.
it looks like the firmware used to have an interface for figuring out what
the device was plugged into, and what it had negotiated, but the newer fw
just has a bit saying "link up" or "link down".

so all we will use the ifmedia stuff for is to show if the link is up or
down.