Allow allocations larger than PAGE_SIZE from pools.
This is solved by special allocators and an obfuscated compare function
for the page header splay tree and some other minor adjustments.

At this moment, the allocator will be picked automagically by pool_init
and you can get a kernel_map allocator if you specify PR_WAITOK in flags
(XXX), default is kmem_map. This will be changed in the future once the
allocator code is slightly reworked. But people want to use it now.

"nag nag nag nag" dlg@