Remove the BUS_DMA_NOCACHE flag from all azalia dma mappings.
The fact that this has only just been implemented, and that this
hardware is only found on x86 shows that this flag should not be needed
for operation of this hardware. It's been here since the initial netbsd
commit the code came from.

"sounds good to me" jakemsr@, ok kettenis@.