Disable all IGPGART chipsets in radeondrm for release.
The RS480 is just fucking evil, the RS690 works ONCE but doesn't on
reboot, and i've never heard about a RS400 being tested, so disable that
too for safety.

This is so that only working chipsets make release, if I just removed
your working chipset, please let me know.

deraadt@ requested that I do this now, not later.