This fixes a case where we could panic on a null deref with a bad vnode in nfs_inactive, on a reboot.
The core of the problem was in nfs_nget, when we lose the race to put a new
nfsnode in the tree, we have previously allocated a vnode, which getnewvnode
has done an insmntque into the nfs mp's mntlist. The problem being we then
try again with a new vnode, abandoning this one on the mntlist, leaving
junk there for us to die on when we unmount.

analysis and bogus fix by art@, correct fix by me. much painful testing by
krw@

ok oga@, art@