Install mandoc_char(7), a comprehensive overview of mandoc(1) special characters and predefined strings.
In kristaps@'s words:
"Note that mandoc_char(7) is a superset of the latest groff_char(7)
(which defines most special characters), groff(7) (more special
characters, mostly formatting), and mdoc_samples(7) (predefined
strings and other nits).  Thus, some of the characters that mandoc(1)
accepts are not accepted by the groff in OpenBSD base."

This is useful because you now have everything in one place, and
in a form authoritative for mandoc(1).  At least as long as we still
have groff(1) in base, we must keep groff_char(7) as well, of course.

"sure" jmc@