update the inteldrm GEM interface to the newer version.
This adds a couple of new ioctls and remove the I915_GEM_EXECBUFFER
ioctl in exchange for EXECBUFFER2 which supports better tiling
parameters. the original interface is removed since our GEM support will
not support it. This temporarily breaks the build of libdrm_intel, this
will be fixed in the next commit.

discussed with matthieu@