It is about time that we stopped pretending simple_locks are locks.
replace ntfs_nthash_slock usage with comments prefixed XXXLOCKING (for
grepability).

This lock looks to be correct, but it could well be the bad way to do it
(having a rwlock for inserts to avoid races inserting the same inode but
then simple locking on list accesses).

approach discussed with deraadt@