*really* filter symbols on elf systems: introduce a new "export_symbols" method that does all the work to produce the right options (one option, so that the *.ver file contains all useful information).
Have to cater to regexp: in particular, make sure to nm on every object
and static library we might need, so we don't lose anything.

Allow for the result to be empty (since some projects use -export-symbols-regex
everywhere, *including on binaries with empty symbol lists*)

should now be clean portswise.

As usual, a non working option means lots of weird fuck-ups to fix first...

thx to millert@ for the hint, aja,jasper,landry for comments and testing.