when handling puts from a client (ie, tftpd is writing a file to disk), we maintain the client state after we've finished writing the file in case the client loses our ack of the last write. unfortunately we didnt close the file we'd just written when we knew it was finished, but only after we clean up the client state after that wait.
because we use FILE stuff to write the file out, its likely some
io flushed to disk until we finish that wait and close the file as
part of cleaning up the client state. if something is coordinating
a bunch of uploads and expects the file to be there after the client
is happy its there, this can be "not good".

this closes the file after we know its finished before proceeding
to hang to handle lost acks to the client.

found by and ok henning@