Fix explicit_bzero regress for Solaris and OS X compatibility
Solaris and OS X clobber the signal stack when returning to the main
stack, which caused the original testing strategy (inspecting the
signal stack once we're back on the main stack) to fail.

To be compatible with this behavior, the regress test now inspects the
signal stack space while we're still executing on it.  This is a bit
iffy because we might clobber it ourselves while inspecting it, but we
as long as its not completely clobbered we should be okay.

thx bcook for the Solaris test account