We cannot easily control the order of the QUERY_STRING keys generated by the search form, it's just the order of the fields in the form. Actually, that's not too bad; the generated URI resembles the generating form.
To minimize confusion for people looking at URIs, give the keys
in the same order when generating URIs for search listings and
search redirections, the latter being used instead of search
listings that would have only one single entry.  Also, if the
manpath is the default, remove it form the generated URIs.