Enable timeouts, just in case(tm).
Even if it's very handy to know where a thread is sleeping in order to
debug HC drivers, users might not like to have to restart their machine
if a transfer timed and nothing will wakeup the discovery thread.

Note that I still haven't seen any hardware timeout in all my tests.