Did you ever wonder why loopback's ifas have a destination address?
It is of course not to make your life more complicated when you are
dealing with ifa_ifwithaddr()!  It was to reuse the point-to-point
code to add a route to 127.0.0.1.  But now we have local routes and
we don't need this hack anymore :)

ok mikeb@ as part of a larger diff.