pool_chk_page iterates over a pages free item lists and checks that the items address is within the page. it does that by masking the item address with the page mask and comparing that to the page address.
however, if we're using large pages with external page headers, we
dont request that the large page be aligned to its size. eg, on an
arch with 4k pages, an 8k large page could be aligned to 4k, so
masking bits to get the page address wont work.

these incorrect checks were distracting while i was debugging large
pages on landisk.

this changes it to do range checks to see if the item is within the
page. it also checks if the item is on the page before checking if
its magic values or poison is right.

ok miod@